Struct noodles_bgzf::writer::Writer

pub struct Writer<W> where
    W: Write, { /* private fields */ }

A BZGF writer.

This implements std::io::Write, consuming uncompressed data and emitting compressed data.

Examples

use noodles_bgzf as bgzf;

let mut writer = bgzf::Writer::new(Vec::new());
writer.write_all(b"noodles-bgzf")?;

let data = writer.finish()?;

Implementations

impl<W> Writer<W> where
W: Write,

pub fn builder(inner: W) -> Builder<W>

Creates a BGZF writer builder.

Examples

use noodles_bgzf as bgzf;
let builder = bgzf::Writer::builder(Vec::new());
let writer = builder.build();

pub fn new(inner: W) -> Self

Creates a writer with a default compression level.

Examples

use noodles_bgzf as bgzf;
let writer = bgzf::Writer::new(Vec::new());

pub fn get_ref(&self) -> &W

Returns a reference to the underlying writer.

Examples

use noodles_bgzf as bgzf;
let writer = bgzf::Writer::new(Vec::new());
assert!(writer.get_ref().is_empty());

pub fn into_inner(self) -> W

Returns the underlying writer.

Examples

use noodles_bgzf as bgzf;
let writer = bgzf::Writer::new(Vec::new());
assert!(writer.into_inner().is_empty());

pub fn position(&self) -> u64

Returns the current position of the stream.

Examples

use noodles_bgzf as bgzf;
let writer = bgzf::Writer::new(Vec::new());
assert_eq!(writer.position(), 0);

pub fn virtual_position(&self) -> VirtualPosition

Returns the current virtual position of the stream.

Panics

This panics if the stream flushed >= 256 TiB of compressed data.

Examples

use noodles_bgzf as bgzf;
let writer = bgzf::Writer::new(Vec::new());
assert_eq!(writer.virtual_position(), bgzf::VirtualPosition::from(0));

pub fn try_finish(&mut self) -> Result<()>

Attempts to finish the output stream by flushing any remaining buffers.

This then appends the final BGZF EOF block.

Examples

use noodles_bgzf as bgzf;

let mut writer = bgzf::Writer::new(Vec::new());
writer.write_all(b"noodles-bgzf")?;

writer.try_finish()?;

pub fn finish(self) -> Result<W>

Returns the underlying writer after finishing the output stream.

This method can only be called once. Any further usage of the writer may result in a panic.

Examples

use noodles_bgzf as bgzf;

let mut writer = bgzf::Writer::new(Vec::new());
writer.write_all(b"noodles-bgzf")?;

let data = writer.finish()?;